+++ android/app/build.gradle
@@ -75,9 +75,9 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.lendsqrfpnews"
+    namespace "com.lendsqrfpnews.newsapp"
     defaultConfig {
-        applicationId "com.lendsqrfpnews"
+        applicationId "com.lendsqrfpnews.newsapp"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
